fix(product): 상품 관련 코드리뷰 반영 (#7)
- 결제 금액 타입을 int -> BigDecimal로 변경
- 중복 제거를 위해 ProductResponse를 record 변환
- 불필요한 코드 정리 및 리팩토링

+++ src/main/java/com/sellon/product/dto/ProductRequest.java
@@ -6,6 +6,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class ProductRequest {
     @NotBlank(message = "상품명은 필수입니다.")
@@ -15,7 +17,7 @@ public class ProductRequest {
 
     @NotNull(message = "가격은 필수입니다.")
     @Min(value = 0, message = "가격은 0 이상이어야 합니다.")
-    private Integer price;
+    private BigDecimal price;
 
     @NotNull(message = "재고는 필수입니다.")
     @Min(value = 0, message = "재고 수량은 0 이상이어야 합니다.")
